requirements.txt
PIL==1.1.7
django==1.5.1
django-cms==2.4.2
south==0.7.2
psycopg2==2.4.2
gunicorn==0.17.2
dj-database-url==0.2.1
django-debug-toolbar==0.9.4
coverage==3.5.1
pylint==0.25.1
fabric==1.6.0
paramiko==1.10.1
pycrypto==2.6.0
Pillow==2.0.0
python-memcached==1.48
MySQL-python==1.2.4
South==0.7.6
django_compressor==1.3
django-appconf==0.6
six==1.3.0
BeautifulSoup==3.2.1
django-blog-zinnia==0.12.3
django-admin-sortable==1.4.9
easy-thumbnails==1.2